Rename files in utils package. Remove deprecated/unused hooks.
