+++ packages/protokit/src/hooks/index.js
@@ -1,2 +1,4 @@
+export * from 'use-enhanced-state';
+
 export * from './useData';
 export * from './useListData';
