+++ packages/utils/src/create-store.js
@@ -1,0 +1,3 @@
+// @todo Deprecate. Use @wp-g2/substate directly.
+// Currently being used in @wp-g2/create-styles
+export * from '@wp-g2/substate';
