--- packages/utils/src/createStore.js
@@ -1,1 +1,0 @@
-export * from '@wp-g2/substate';
