+++ packages/utils/src/deep-equal.js
