+++ packages/utils/src/hooks/use-sealed-state.js
