--- packages/utils/src/hooks/useBoolState.js
@@ -1,1 +1,0 @@
-export { useBoolState } from 'use-enhanced-state';
