--- packages/utils/src/hooks/useListState.js
@@ -1,1 +1,0 @@
-export { useListState } from 'use-enhanced-state';
