+++ packages/utils/src/shallow-equal.js
