+++ packages/utils/src/simple-equal.js
